+++ .github/actions/wasm-ci/action.yml
@@ -9,10 +9,6 @@ runs:
       working-directory: ${{github.workspace}}/source/webassembly
       run: npm pack
 
-    - name: Run Xvfb
-      shell: bash
-      run: Xvfb $DISPLAY -screen 0 1280x1024x24 &
-
     - name: Test
       shell: bash
       working-directory: ${{github.workspace}}/source/webassembly
