+++ .github/workflows/ci.yml
@@ -977,7 +977,6 @@ jobs:
 
     env:
       CMAKE_BUILD_PARALLEL_LEVEL: 4
-      DISPLAY: :0
       F3D_DOCKER_TIMESTAMP: ${{needs.default_versions.outputs.timestamp}}
 
     steps:
