+++ doc/dev/BUILD.md
@@ -1,5 +1,8 @@
 # Build guide
 
+> [!NOTE]
+> For WebAssembly cross-compilation, follow the dedicated guide [here](./BUILD_WASM.md).
+
 F3D uses a CMake based build system, so building F3D just requires installing
 needed dependencies, configuring and building. If you are not used to such processes
 please take a look at our [getting started guide](GETTING_STARTED.md).
