+++ doc/libf3d/LANGUAGE_BINDINGS.md
@@ -58,4 +58,67 @@ public class F3DExample {
 
 If the Javascript bindings have been generated by building F3D with webassembly and emscriptem, the libf3d can be used directly from a browser.
 
-See the dedicated [build](../dev/BUILD_WASM.md) guide, this [example app](https://github.com/f3d-app/f3d/blob/master/webassembly/example.html) which you can test live [here](https://f3d.app/web/)!
+See this [example app](https://github.com/f3d-app/f3d/blob/master/example/libf3d/web) which you can test live [here](https://f3d.app/web/)!
+
+The package can be installed using npm, by running the following command:
+
+```bash
+npm install f3d
+```
+
+Create the target DOM element `<canvas id="canvas"></canvas>` in your app, then use the following javascript code:
+
+```javascript
+// import the module
+import f3d from "f3d";
+
+// setup a setting object
+const settings = {
+  canvas: document.getElementById("canvas"),
+  setupOptions: (options) => {
+    // background must be set to black for proper blending with transparent canvas
+    options.set_color("render.background.color", 0, 0, 0);
+
+    // make it look nice
+    options.toggle("render.effect.antialiasing.enable");
+    options.toggle("render.effect.tone_mapping");
+    options.toggle("render.effect.ambient_occlusion");
+    options.toggle("render.hdri.ambient");
+
+    // display widgets
+    options.toggle("ui.axis");
+    options.toggle("render.grid.enable");
+  },
+};
+
+f3d(settings)
+  .then(async (Module) => {
+    // write a 3D file located on the server to the internal filesystem
+    const modelFile = await fetch("/assets/example.obj").then((b) => b.arrayBuffer());
+    Module.FS.writeFile("example.obj", new Uint8Array(modelFile));
+
+    // automatically load all supported file format readers
+    Module.Engine.autoloadPlugins();
+
+    // create an engine
+    Module.engineInstance = Module.Engine.create();
+
+    Module.setupOptions(Module.engineInstance.getOptions());
+
+    // setup the window size based on the canvas size
+    const scale = window.devicePixelRatio;
+    Module.engineInstance
+      .getWindow()
+      .setSize(Module.canvas.clientWidth, scale * Module.canvas.clientHeight);
+
+    const scene = Module.engineInstance.getScene();
+    scene.add("/example.obj");
+
+    // do a first render and start the interactor
+    Module.engineInstance.getWindow().render();
+    Module.engineInstance.getInteractor().start();
+  })
+  .catch((error) => {
+    console.error("Internal exception: " + error);
+  });
+```
